requirements.txt
# Core AI APIs
openai>=1.0.0
anthropic>=0.7.0
google-generativeai>=0.3.0
gspread>=5.7.0
google-auth>=2.0.0
google-auth-oauthlib>=0.8.0
google-auth-httplib2>=0.1.0
requests>=2.25.0
beautifulsoup4>=4.9.0
lxml>=4.6.0
tqdm>=4.60.0
python-dotenv>=0.19.0

# Data processing
pandas>=1.3.0
numpy>=1.21.0
jsonschema>=3.2.0

# XML processing for NCBI
xmltodict>=0.12.0

# Optional: for better JSON handling
ujson>=5.0.0

# Development tools (optional)
pytest>=6.0.0
black>=21.0.0
flake8>=3.8.0